refactor plan view
